fixed int

+++ FVCalc.java
@@ -11,6 +11,6 @@ public class FVCalc {
 	    double resultPow= (Math.pow(1+ rate1,n));
 		//System.out.println(resultPow);
 		double futureValue= (currentValue*resultPow);
-		System.out.println("After " + n +" years, " +currentValue+ "$ saved at "+rate +"% will yield " + futureValue);
+		System.out.println("After " + n +" years, " +currentValue+ "$ saved at "+rate +"% will yield " + ((int)futureValue));
 	}
 }
